ai_grader_project/requirements.txt
torch
transformers
pillow
sentence-transformers
scikit-learn
opencv-python
flask
firebase-admin
gunicorn
